pyproject.toml
[project]
name = "bank_payment_parser"
authors = [
    {name = "Frappe", email = "support@frappe.io"}
]
description = "Scalable bank payment advice PDF parser with customer-specific parsing"
requires-python = ">=3.10"
readme = "README.md"
dynamic = ["version"]
dependencies = [
    "pdfminer.six>=20221105",
    # "frappe~=15.0.0" # Installed and managed by bench.
]

[build-system]
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

# These dependencies are only installed when developer mode is enabled
[tool.bench.dev-dependencies]
# package_name = "~=1.1.0"

[project.optional-dependencies]
ocr = [
    "pytesseract>=0.3.10",
    "pdf2image>=1.16.3",
    "Pillow>=10.0.0",
]

[tool.ruff]
line-length = 110
target-version = "py310"

[tool.ruff.lint]
select = [
    "F",
    "E",
    "W",
    "I",
    "UP",
    "B",
    "RUF",
]
ignore = [
    "B017",
    "B018",
    "B023",
    "B904",
    "E101",
    "E402",
    "E501",
    "E741",
    "F401",
    "F403",
    "F405",
    "F722",
    "W191",
]
typing-modules = ["frappe.types.DF"]

[tool.ruff.format]
quote-style = "double"
indent-style = "tab"
docstring-code-format = true
